+++ src/Infrastructure/CleanArc.Infrastructure.Identity/Identity/Store/AppUserStore.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 
 namespace CleanArc.Infrastructure.Identity.Identity.Store;
 
-public class AppUserStore:UserStore<User,Role,ApplicationDbContext,int,UserClaim,UserRole,UserLogin,UserToken,RoleClaim>
+public class AppUserStore:UserStore<User,Role,ApplicationDbContext,long,UserClaim,UserRole,UserLogin,UserToken,RoleClaim>
 {
     public AppUserStore(ApplicationDbContext context, IdentityErrorDescriber describer = null) : base(context, describer)
     {
